Delete todo confirmation added, after delete todo refresh page.

+++ TODO app/components/Todo/DeleteTodo.js
@@ -1,9 +1,13 @@
 "use client"
-
+import { deleteTodo } from "@/lib/todoServerAction"
+import { useRouter } from "next/navigation"
 import { MdDeleteForever } from "react-icons/md"
 
 export default function DeleteTodo({ id }) {
+	const router = useRouter()
 	async function handleDelete() {
+		let confirmed = confirm("Do you want to delte todo!")
+		if (!confirmed) return
 		await fetch(`/api/todo?id=${id}`, {
 			method: "DELETE"
 		})
@@ -14,6 +18,10 @@ export default function DeleteTodo({ id }) {
 				_deleteRes.call === 1
 					? alert("Delete successful")
 					: alert("Not deleted, Something went wrong")
+
+				if (_deleteRes.call === 1) {
+					router.refresh()
+				}
 			})
 	}
